Update to highlight.js fork tower-enhancements branch commit 163c6934:
Split out handling of Objective-C @keywords, to fix highlighting skipping the leading ‘@‘ of keywords like “@property”, and treaing the remainder as a non-keyword identifier, due I think to ‘@‘ being treated as a boundary character rather than a valid keyword character. (‘@‘ is not a member of the \w regex character class).